agent update refactor
