+++ makepkg.py
@@ -9,7 +9,7 @@ import sys
 
 from azurelinuxagent.common.version import AGENT_NAME, AGENT_VERSION, \
     AGENT_LONG_VERSION
-from azurelinuxagent.ga.update import AGENT_MANIFEST_FILE
+from azurelinuxagent.ga.agent_update import AGENT_MANIFEST_FILE
 
 MANIFEST = '''[{{
     "name": "{0}",
